+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### 🛑 Breaking changes 🛑
 
+- Remove deprecated pdata funcs/structs from v0.50.0 (#5345)
+
 ### 🚩 Deprecations 🚩
 
 ### 💡 Enhancements 💡